requirements.txt
flask
Flask-Bootstrap
flask_sqlalchemy
python-dotenv